I fixed the links to all of the webpages.
Use this method from now on.

I also added some text to the pages so that they represent the information that is supposed to be shown with headers and such.

I plan on working more on the actual content that is on each page including styling and spacing tonight.

Just wanted you to have this so we can work on the same pages.

P.S. if you update this and continue, it should work considering your current file path. just name them like I did.

<3